+++ src/Trade360SDK.Feed/Configuration/RmqConnectionSettings.cs
@@ -6,6 +6,7 @@
         public int Port { get; set; }
         public string? VirtualHost { get; set; }
         public int PackageId { get; set; }
+        public string? CustomQueueName { get; set; }
         public string? UserName { get; set; }
         public string? Password { get; set; }
         public ushort PrefetchCount { get; set; } = 100; // Default 100;
@@ -14,6 +15,7 @@
         public bool AutoAck { get; set; } = true; // Default true;
         public int RequestedHeartbeatSeconds { get; set; } = 30; // Default 30 seconds
         public int NetworkRecoveryInterval { get; set; } = 30; // Default 30 seconds
+        public bool SslEnabled { get; set; }
     }
 
 }